Reorganize code, prototype code for auto iteration
